use ListView.separated with ItemListDivider and move BudgetItem to own file
