+++ lib/shared/widgets/itemLists/item_list_container.dart
@@ -1,25 +1,17 @@
 import 'package:flutter/material.dart';
 
-/// Container with a item as [child] and a divider below
-class ItemListContainer extends StatelessWidget {
-  final Widget child;
-
-  const ItemListContainer({
-    required this.child,
+///
+class ItemListDivider extends StatelessWidget {
+  const ItemListDivider({
     super.key,
   });
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        child,
-        const Divider(
-          thickness: 1,
-          indent: 10,
-          endIndent: 10,
-        ),
-      ],
+    return const Divider(
+      thickness: 1,
+      indent: 10,
+      endIndent: 10,
     );
   }
 }
